requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==1.10.7
yfinance==0.1.63
pandas==1.3.3
numpy==1.21.2
fredapi==0.4.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.19.1
openai==0.28.0
python-dotenv==0.21.0
scikit-learn==0.24.2
joblib==1.0.1
streamlit==1.22.0
plotly==5.3.1
pytest==7.3.1
requests-mock==1.10.0
